chore: format code

+++ contracts/lending/AErc20Immutable.sol
@@ -20,14 +20,16 @@ contract AErc20Immutable is AErc20 {
      * @param decimals_ ERC-20 decimal precision of this token
      * @param admin_ Address of the administrator of this token
      */
-    constructor(address underlying_,
-                AvatrollerInterface comptroller_,
-                InterestRateModel interestRateModel_,
-                uint initialExchangeRateMantissa_,
-                string memory name_,
-                string memory symbol_,
-                uint8 decimals_,
-                address payable admin_) {
+    constructor(
+        address underlying_,
+        AvatrollerInterface comptroller_,
+        InterestRateModel interestRateModel_,
+        uint256 initialExchangeRateMantissa_,
+        string memory name_,
+        string memory symbol_,
+        uint8 decimals_,
+        address payable admin_
+    ) {
         // Creator of the contract is admin during initialization
         admin = payable(msg.sender);
